==> 0_train_val_test_split.ipynb <==
import numpy as np
import pandas as pd

# pull prior split
oldsplit = pd.read_csv('d:/Dropbox (Partners Healthcare)/profile_9-2022/derived_data/split_9-2022.csv')

oldsplit.head()
oldsplit.dfci_mrn.nunique()

new_dataset = pd.read_csv('d:/Dropbox (Partners Healthcare)/profile_3-2023/structured_data/REQ_KK71_105304_2_PT_INFO_STATUS_REGISTRATION.csv', low_memory=False)[['DFCI_MRN']].rename(columns={'DFCI_MRN':'dfci_mrn'})
new_dataset.info()

new_dataset = new_dataset.groupby('dfci_mrn').first().reset_index()
new_dataset.shape[0]

new_patients = new_dataset[~new_dataset.dfci_mrn.isin(oldsplit.dfci_mrn)]

new_patients.shape[0]


new_true_test_mrns = new_patients.sample(frac=0.1, random_state=1)
remainder_mrns = new_patients[new_patients['dfci_mrn'].isin(new_true_test_mrns.dfci_mrn) == False]
new_validation_mrns = remainder_mrns.sample(frac=0.111111, random_state=2)
new_training_mrns = remainder_mrns[remainder_mrns['dfci_mrn'].isin(new_validation_mrns.dfci_mrn) == False]

new_true_test_mrns['split'] = 'test'
new_validation_mrns['split'] = 'validation'
new_training_mrns['split'] = 'train'

new_split = pd.concat([new_training_mrns, new_validation_mrns, new_true_test_mrns], axis=0)

new_split.dfci_mrn.isin(oldsplit.dfci_mrn).value_counts()

output = pd.concat([oldsplit[['dfci_mrn','split']], new_split], axis=0).sort_values(by='dfci_mrn').reset_index(drop=True)

output.info()

output.to_csv('d:/Dropbox (Partners Healthcare)/profile_3-2023/derived_data/split_3-2023.csv')

oldsplit.info()

output.info()

output.split.value_counts()
