==> 1_json_to_csv_imaging.ipynb <==
# notebook used for transforming imaging reports in JSON format into CSV file

import numpy as np
import pandas as pd
import json
import glob

prefix = '/mnt/d/Dropbox (Partners Healthcare)/profile_3-2023/'


def parse_imaging_json(file_name):
    with open(file_name, encoding='utf-8') as json_file:
        imaging_dict = json.load(json_file)
    imaging_frame = pd.DataFrame.from_records(imaging_dict['response']['docs'])   
    imaging_frame = imaging_frame[['DFCI_MRN','EVENT_DATE','RPT_TEXT','IMPRESSION_TEXT','PROC_DESC']]
    imaging_frame = imaging_frame.rename(columns={'DFCI_MRN':'dfci_mrn', 'EVENT_DATE':'date', 'RPT_TEXT':'text', 'IMPRESSION_TEXT':'impression_text','PROC_DESC':'scan_type'})
    
    imaging_frame.date = pd.to_datetime(imaging_frame.date.str[:10])
    
    imaging_frame.text = imaging_frame.text.str.replace('\n|\r', ' ')
    imaging_frame.text = imaging_frame.text.str.encode('ascii', 'ignore')
    imaging_frame.text = [str(x, 'ascii') for x in imaging_frame.text]
    imaging_frame.text = imaging_frame.text.str.replace('20\d\d',' ')

    imaging_frame['text'] = imaging_frame.text.str.replace("\\s+", " ", regex=True)


    #imaging_frame.text = imaging_frame.text.str.lower()
    
    imaging_frame = imaging_frame[~imaging_frame.scan_type.str.contains('XR ')]
    
    imaging_frame = imaging_frame[~imaging_frame.scan_type.str.contains('PACS')]

    
    return imaging_frame
        

filename_list = [x for x in glob.glob(prefix+'/text/imaging/*Imag*.json')]
filename_list

report_list = [parse_imaging_json(x) for x in filename_list]

report_frame = pd.concat(report_list, axis=0)

report_frame.info()

registration = pd.read_csv(prefix+'structured_data/REQ_KK71_105304_2_PT_INFO_STATUS_REGISTRATION.csv', sep=',', encoding='iso-8859-1', low_memory=False)[['DFCI_MRN', 'PATIENT_ID', 'HYBRID_DEATH_IND', 'HYBRID_DEATH_DT']].rename(columns={'DFCI_MRN':'dfci_mrn', 'PATIENT_ID':'patient_id', 'HYBRID_DEATH_IND':'hybrid_death_ind', 'HYBRID_DEATH_DT':'hybrid_death_dt'})
registration.info()

report_frame['dfci_mrn'] = pd.to_numeric(report_frame.dfci_mrn)
report_frame = pd.merge(report_frame, registration, on='dfci_mrn')


# pull tumor sequencing data to get cancer diagnosis per Oncotree and NGS date
specimens = pd.read_csv(prefix+'structured_data/REQ_KK71_105304_2_GENOMIC_SPECIMEN.csv', encoding='ISO-8859-1', low_memory=False)
specimens = specimens[['PATIENT_ID','PRIMARY_CANCER_DIAGNOSIS', 'TEST_ORDER_DT']].groupby('PATIENT_ID').first().reset_index().rename(columns={'PATIENT_ID':'patient_id', 'PRIMARY_CANCER_DIAGNOSIS':'primary_cancer_diagnosis', 'TEST_ORDER_DT':'genomics_date'})
specimens = pd.merge(specimens, registration, on='patient_id')
specimens = specimens[~specimens.primary_cancer_diagnosis.isnull()]
# no leukemia or myeloma
specimens = specimens[~specimens.primary_cancer_diagnosis.str.contains("Leukemia|Myeloma")]
specimens['genomics_date'] = pd.to_datetime(specimens.genomics_date)
specimens = specimens.sort_values(by=['dfci_mrn','genomics_date']).reset_index(drop=True)
specimens = specimens.groupby('dfci_mrn').first().reset_index()
specimens = specimens[['dfci_mrn','primary_cancer_diagnosis', 'genomics_date']]
specimens.primary_cancer_diagnosis.value_counts()

specimens.shape[0]

report_frame = pd.merge(report_frame, specimens, on=['dfci_mrn'])
report_frame['after_profile'] = np.where(report_frame.genomics_date.isnull() | (report_frame.date < report_frame.genomics_date), 0, 1)
report_frame['hybrid_death_dt'] = pd.to_datetime(report_frame.hybrid_death_dt)


# pull train/val/test split
split = pd.read_csv(prefix+'derived_data/split_3-2023.csv')

final_output = pd.merge(report_frame, split[['dfci_mrn', 'split']], on='dfci_mrn')
# no external scans without interpretation
final_output = final_output[~final_output.scan_type.str.contains('NO INTERPRETATION')]
final_output.info()

prefix

final_output.to_csv(prefix+'derived_data/all_imaging_for_prissmm.csv')

final_output = pd.read_csv(prefix+'derived_data/all_imaging_for_prissmm.csv')

final_output.scan_type.value_counts()[0:20]
