==> 7_imaging_demographics.ipynb <==
# this notebook just calculates basic demographics for the DFCI cohort for imaging reports

import numpy as np
import pandas as pd

prefix = '/mnt/d/Dropbox (Partners Healthcare)/'
pt_dataset_list = []
for cancer_type in ['nsclc_phase2_existing','crc','breast','pancreas','bladder','prostate']:
    pts = pd.read_csv(prefix+'profile_3-2023/redcaps/' + cancer_type + '_ptchars.csv', low_memory=False)[['record_id','internal_mrn','genie_patient_id']]
    pts = pts.sort_values(by=['record_id','internal_mrn']).reset_index(drop=True)
    pts = pts.groupby('internal_mrn').first().reset_index(drop=False).rename(columns={'internal_mrn':'dfci_mrn'})
    pts['cancer_type'] = cancer_type
    pts = pts[['dfci_mrn','genie_patient_id','cancer_type']]
    pt_dataset_list.append(pts)


pt_dataset = pd.concat(pt_dataset_list, axis=0)

pt_dataset = pt_dataset.groupby('dfci_mrn').first().reset_index()

pt_dataset.info()

prefix = '/mnt/d/Dropbox (Partners Healthcare)/profile_3-2023/'

specimens = pd.read_csv(prefix+'structured_data/REQ_KK71_105304_2_GENOMIC_SPECIMEN.csv', encoding='ISO-8859-1', low_memory=False)
specimens = specimens[['DFCI_MRN', 'TEST_ORDER_DT']]
specimens['TEST_ORDER_DT'] = pd.to_datetime(specimens.TEST_ORDER_DT)
specimens = specimens.sort_values(by=['DFCI_MRN','TEST_ORDER_DT']).groupby('DFCI_MRN').first().reset_index()


specimens.info()

#no_genie = pt_dataset[pt_dataset.genie_patient_id.isnull()]

#reports[reports.dfci_mrn.isin(no_genie.dfci_mrn)].any_cancer.value_counts()

#no_genie.cancer_type.value_counts()

#pt_dataset.genie_patient_id.nunique()

registration = pd.read_csv(prefix+'structured_data/REQ_KK71_105304_2_PT_INFO_STATUS_REGISTRATION.csv', sep=',', encoding='iso-8859-1', low_memory=False)
registration = registration.groupby('DFCI_MRN').first().reset_index()
registration = registration[['DFCI_MRN','GENDER_NM','BIRTH_DT']]
registration['BIRTH_DT'] = pd.to_datetime(registration.BIRTH_DT)

registration = pd.merge(registration, specimens, on='DFCI_MRN')
registration['age_at_sequencing'] = (registration.TEST_ORDER_DT - registration.BIRTH_DT) / np.timedelta64(1, 'Y')
#registration.head()

demographics = pd.read_csv(prefix+'structured_data/REQ_KK71_105304_2_DEMOGRAPHICS_REGISTRATION.csv', sep=',', encoding='iso-8859-1', low_memory=False)
demographics = demographics.groupby('DFCI_MRN').first().reset_index()
demographics.info()

demographics['race'] = np.select([
    demographics.RACE_NAT_HAWAIIAN_PACIFIC_IND == 'Y',
    demographics.RACE_AM_INDIAN_ALASKAN_IND == 'Y',
    demographics.RACE_BLACK_IND == 'Y',
    demographics.RACE_ASIAN_IND == 'Y',
    demographics.RACE_WHITE_IND == 'Y'],
    ['Native Hawaiian/Pacific Islander',
     'American Indian/Alaskan',
     'Black',
     'Asian',
     'White'],
     'Other/unknown')

demographics = demographics[['DFCI_MRN','race','HISPANIC_IND']]

demographics = pd.merge(demographics, registration, on='DFCI_MRN')

demographics.GENDER_NM.value_counts()

demographics = demographics.rename(columns={'DFCI_MRN':'dfci_mrn'})

#demographics.head()

demographics['agecat'] = np.select([
    demographics.age_at_sequencing < 30,
    demographics.age_at_sequencing < 40,
    demographics.age_at_sequencing < 50,
    demographics.age_at_sequencing < 60,
    demographics.age_at_sequencing < 70,
    demographics.age_at_sequencing < 80,
    True
],
    ['< 30',
     '30-39',
     '40-49',
     '50-59',
     '60-69',
     '70-79',
     '80+'])

reports = pd.read_csv('/mnt/d/Dropbox (Partners HealthCare)/profile_3-2023/derived_data/labeled_imaging_prissmm.csv')

reports = pd.merge(reports, demographics, on='dfci_mrn')

# mixed response counts as progression
reports['progression'] = np.where(reports.class_status==3,1,reports.progression)


reports.info()

reports.image_scan_type.value_counts()

report_demographics = reports.copy()
report_demographics.shape

report_demographics.groupby('dfci_mrn').first().shape

for var in ['GENDER_NM', 'race', 'HISPANIC_IND','agecat', 'cancer_type','any_cancer','progression','response','brain_met','bone_met','adrenal_met','liver_met','lung_met','node_met','peritoneal_met']:
    print('document level')
    print(var)
    print(report_demographics[var].value_counts())
    print(report_demographics[var].value_counts()/report_demographics.shape[0])
    print("\n")

patient_demographics = reports.groupby('dfci_mrn').first().copy()

for var in ['GENDER_NM', 'race', 'HISPANIC_IND','agecat', 'cancer_type','any_cancer','progression','response','brain_met','bone_met','adrenal_met','liver_met','lung_met','node_met','peritoneal_met']:
    print('patient level')
    print(var)
    print(patient_demographics[var].value_counts())
    print(patient_demographics[var].value_counts()/patient_demographics.shape[0])
    print("\n")
